search_api/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==11.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
